+++ sample/Source/SentryPlayground/SentryPlaygroundUtils.h
@@ -15,6 +15,7 @@ enum class ESentryAppTerminationType : uint8
 	InvalidMemoryAccess,
 	StackOverflow,
 	OutOfMemory,
+	MemoryCorruption,
 	FastFail,
 	RenderThreadCrash,
 	GpuDebugCrash,